Add production of data packets from expired flows.
We build up a basic template using some of the fields we're collecting.
From the template we build up data packets. The data packets instead
of being a large chunk of data being written to are structs which contain
some vectors for the various sets.

It would currently require a function to take in a pointer to memory and
write out the raw packet. I see we have two ways we can proceed:

- Keep the structure we have of VPP vectors. Create a function to transform
  it into a raw packet to be written out on the tx buffer.

- Change it so we have simply have a chunk of memory on the heap with all
  the data packets pre-propared.

The code only generates Netflow V10 packets, not V9.