Improve netflow V10 structs to mirror real data.
This moves away from using a VPP vector to a huge chunk of memory on
heap, to access the fields one needs to walk over the fields using
the offsets in the template. One of the benefits of this method is
the netflow V10 structs looks a similiar to what is sent on the line
meaning you can cast the buffer using these structs similiar to the
other structs in VPP.

This also adds a method which takes a vlib_buffer_t and a packet
and will write this data to the buffer. This should be checked
and if nessesery refined and/or fixed when Asumu's half lands.